requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
Django==5.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pillow==10.4.0
preshed==3.0.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytesseract==0.3.13
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
setuptools==73.0.1
shellingham==1.5.4
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
textblob==0.18.0.post0
thinc==8.2.5
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
